+++ cpp/subprojects/seco/src/mlrl/seco/heuristics/heuristic_m_estimate.cpp
@@ -12,27 +12,27 @@ namespace seco {
     class MEstimate final : public IHeuristic {
         private:
 
-            const float64 m_;
+            const float32 m_;
 
         public:
 
             /**
              * @param m The value of the "m" parameter. Must be at least 0
              */
-            MEstimate(float64 m) : m_(m) {}
+            MEstimate(float32 m) : m_(m) {}
 
-            float64 evaluateConfusionMatrix(float64 cin, float64 cip, float64 crn, float64 crp, float64 uin,
-                                            float64 uip, float64 urn, float64 urp) const override {
+            float32 evaluateConfusionMatrix(float32 cin, float32 cip, float32 crn, float32 crp, float32 uin,
+                                            float32 uip, float32 urn, float32 urp) const override {
                 if (std::isinf(m_)) {
                     // Equivalent to weighted relative accuracy
                     return wra(cin, cip, crn, crp, uin, uip, urn, urp);
                 } else if (m_ > 0) {
                     // Trade-off between precision and weighted relative accuracy
-                    float64 numCoveredEqual = cin + crp;
-                    float64 numCovered = numCoveredEqual + cip + crn;
-                    float64 numUncoveredEqual = uin + urp;
-                    float64 numEqual = numCoveredEqual + numUncoveredEqual;
-                    float64 numTotal = numCovered + numUncoveredEqual + uip + urn;
+                    float32 numCoveredEqual = cin + crp;
+                    float32 numCovered = numCoveredEqual + cip + crn;
+                    float32 numUncoveredEqual = uin + urp;
+                    float32 numEqual = numCoveredEqual + numUncoveredEqual;
+                    float32 numTotal = numCovered + numUncoveredEqual + uip + urn;
 
                     if (numTotal > 0) {
                         return (numCoveredEqual + (m_ * (numEqual / numTotal))) / (numCovered + m_);
@@ -54,28 +54,28 @@ namespace seco {
     class MEstimateFactory final : public IHeuristicFactory {
         private:
 
-            const float64 m_;
+            const float32 m_;
 
         public:
 
             /**
              * @param The value of the "m" parameter. Must be at least 0
              */
-            MEstimateFactory(float64 m) : m_(m) {}
+            MEstimateFactory(float32 m) : m_(m) {}
 
             std::unique_ptr<IHeuristic> create() const override {
                 return std::make_unique<MEstimate>(m_);
             }
     };
 
-    MEstimateConfig::MEstimateConfig() : m_(22.466) {}
+    MEstimateConfig::MEstimateConfig() : m_(22.466f) {}
 
-    float64 MEstimateConfig::getM() const {
+    float32 MEstimateConfig::getM() const {
         return m_;
     }
 
-    IMEstimateConfig& MEstimateConfig::setM(float64 m) {
-        util::assertGreaterOrEqual<float64>("m", m, 0);
+    IMEstimateConfig& MEstimateConfig::setM(float32 m) {
+        util::assertGreaterOrEqual<float32>("m", m, 0);
         m_ = m;
         return *this;
     }
